+++ Marlin/src/inc/MarlinConfigPre.h
@@ -31,7 +31,7 @@
 #include "../core/boards.h"
 #include "../core/macros.h"
 #include "../core/millis_t.h"
-#include "Version.h"
+#include "../../Version.h"
 #include "../../Configuration.h"
 #include "Conditionals_LCD.h"
 #include "../core/drivers.h"
